requirements.txt
python-telegram-bot==20.7
selenium==4.17.2
requests==2.31.0